# Documentation/Swagger/swagger-rental-of-vehicles/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Rental of Vehicles
  description: "This API was created with the aim of managing a dealership's rental\
    \ registration, storing data and relationships, such as vehicles and their branches\
    \ and users and their rentals."
  version: "1.0"
servers:
- url: http://localhost:8000/
  description: Sandbox server (uses test data)
paths:
...
  /rents:
    get:
      tags:
      - Rents
      summary: Return a list of all rents
      operationId: rents_get
      responses:
        "200":
          description: An object containing a JSON array of rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - count: 1
                results:
                - id: 1
                  vehicle: "00314270183"
                  staff: "10452749417"
                  client: "00928095371"
                  status: A
                  outlet_branch: 1
                  appointment_date: 2024-02-27
                  requested_days: 5
                  actual_days: 0
                  rent_deposit: 150
                  daily_cost: 14.59
                  additional_daily_cost: 0.18
                  driver:
                  - "00928095371"
                  additional_items:
                  - id: 1
                    number: 1
                    rental: 1
                    additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rents/{id}:
    get:
      tags:
      - Rents
      summary: Return an object containing rent information
      operationId: rents_id_get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxItems: 1
          type: string
        example: 1
      responses:
        "200":
          description: An object that represents a rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - id: 1
                vehicle: "00314270183"
                staff: "10452749417"
                client: "00928095371"
                status: A
                outlet_branch: 1
                appointment_date: 2024-02-27
                requested_days: 5
                actual_days: 0
                rent_deposit: 150
                daily_cost: 14.59
                additional_daily_cost: 0.18
                driver:
                - "00928095371"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rent/{id}:
    put:
      tags:
      - Rents
      summary: Update a rent and return an object with the new values
      operationId: rent_id_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxItems: 1
          type: string
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/rent_id_body'
        required: true
      responses:
        "200":
          description: An object informing the new rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - vehicle: "00314270183"
                client: "00928095371"
                status: L
                requested_days: 5
                rent_deposit: 150
                driver:
                - "00928095371"
                - "01044008628"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /appointment:
    post:
      tags:
      - Rents
      summary: Register a new appointment
      operationId: appointment_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/appointment_body'
        required: true
      responses:
        "201":
          description: An object that represents an appointment
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - vehicle: "00314270183"
                staff: "10452749417"
                client: "00928095371"
                status: A
                appointment_date: 2024-02-27
                requested_days: 5
                rent_deposit: 150
                driver:
                - "00928095371"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /appointment/{id}:
    put:
      tags:
      - Rents
      summary: Update an appointment and return an object with the new values
      operationId: appointment_id_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxItems: 1
          type: string
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/appointment_id_body'
        required: true
      responses:
        "200":
          description: An object informing the new rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - vehicle: "00314270183"
                staff: "10452749417"
                client: "00928095371"
                status: A
                appointment_date: 2024-02-27
                requested_days: 5
                rent_deposit: 150
                driver:
                - "00928095371"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rents/{id}/appointment_to_rent:
    put:
      tags:
      - Rents
      summary: Update the status of a rental from appointment to rented and return
        an object with the new values
      operationId: rents_id_appointment_to_rent_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxItems: 1
          type: string
        example: 1
      responses:
        "200":
          description: An object informing the new rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - id: 1
                vehicle: "00314270183"
                staff: "10452749417"
                client: "00928095371"
                status: L
                outlet_branch: 16
                appointment_date: 2024-02-27
                rent_date: 2024-02-27
                requested_days: 5
                rent_deposit: 150
                daily_cost: 14.59
                additional_daily_cost: 0.18
                driver:
                - "00928095371"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rents/{id}/vehicle_devolution:
    put:
      tags:
      - Rents
      summary: Update the status of a rental from rented to returned and return an
        object with the new values
      operationId: rents_id_vehicle_devolution_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxItems: 1
          type: string
        example: 1
      responses:
        "200":
          description: An object informing the new rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - id: 1
                vehicle: "00314270183"
                staff: "10452749417"
                client: "00928095371"
                status: D
                outlet_branch: 1
                arrival_branch: 1
                distance_branch: 0
                rent_date: 2024-02-08
                devolution_date: 2024-02-12
                requested_days: 5
                actual_days: 5
                fines: 0
                rent_deposit: 150
                daily_cost: 14.59
                additional_daily_cost: 0.18
                return_rate: 0
                total_cost: 73.85
                driver:
                - "00928095371"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rents/{id}/cancel_appointment:
    put:
      tags:
      - Rents
      summary: Update the status of a rental from appointment to canceled and return
        an object with the new values
      operationId: rents_id_cancel_appointment_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxItems: 1
          type: string
        example: 1
      responses:
        "200":
          description: An object informing the new rent information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - id: 1
                vehicle: "00314270183"
                staff: "10452749417"
                client: "00928095371"
                status: C
                outlet_branch: 16
                appointment_date: 2024-03-01
                requested_days: 7
                actual_days: 0
                rent_deposit: 250
                daily_cost: 14.59
                additional_daily_cost: 0.36
                driver:
                - "00928095371"
                - "01044008628"
                additional_items:
                - id: 1
                  number: 1
                  rental: 1
                  additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rents/expired_appointments:
    get:
      tags:
      - Rents
      summary: Return a list of all expired appointments
      operationId: rents_expired_appointments_get
      responses:
        "200":
          description: An object containing a JSON array of expired appointments information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - count: 1
                results:
                - id: 1
                  vehicle: "00314270183"
                  staff: "10452749417"
                  client: "00928095371"
                  status: A
                  outlet_branch: 1
                  appointment_date: 2024-02-01
                  requested_days: 5
                  actual_days: 0
                  rent_deposit: 150
                  daily_cost: 14.59
                  additional_daily_cost: 0.18
                  driver:
                  - "00928095371"
                  additional_items:
                  - id: 1
                    number: 1
                    rental: 1
                    additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
  /rents/expired_returns:
    get:
      tags:
      - Rents
      summary: Return a list of all expired returns
      operationId: rents_expired_returns_get
      responses:
        "200":
          description: An object containing a JSON array of expired returns information
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
              - count: 1
                results:
                - id: 1
                  vehicle: "00314270183"
                  staff: "10452749417"
                  client: "00928095371"
                  status: L
                  outlet_branch: 1
                  appointment_date: 2024-02-01
                  rent_date: 2024-02-01
                  requested_days: 5
                  actual_days: 0
                  rent_deposit: 150
                  daily_cost: 14.59
                  additional_daily_cost: 0.18
                  driver:
                  - "00928095371"
                  additional_items:
                  - id: 1
                    number: 1
                    rental: 1
                    additional_item: 1
      security:
      - basicAuth: []
      x-openapi-router-controller: swagger_server.controllers.rents_controller
components:
  schemas:
...
    rent_id_body:
      type: object
      properties:
        driver:
          type: array
          items:
            type: string
            example: "[\"00928095371\",\"01044008628\"]"
    appointment_body:
      type: object
      properties:
        vehicle:
          type: string
          example: "00314270183"
        client:
          type: string
          example: "00928095371"
        appointment_date:
          type: string
          example: 2024-02-27
        requested_days:
          type: integer
          example: 5
        rent_deposit:
          type: number
          example: 150
        driver:
          type: array
          items:
            type: string
            example: "[\"00928095371\"]"
        additional_items:
          type: array
          items:
            type: object
            example:
            - additional_item: "1,"
              number: 1
    appointment_id_body:
      type: object
      properties:
        vehicle:
          type: string
          example: "00314270183"
        client:
          type: string
          example: "00928095371"
        appointment_date:
          type: string
          example: 2024-02-27
        requested_days:
          type: integer
          example: 5
        rent_deposit:
          type: number
          example: 150
        driver:
          type: array
          items:
            type: string
            example: "[\"00928095371\"]"
        additional_items:
          type: array
          items:
            type: object
            example:
            - additional_item: "1,"
              number: 1
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_basicAuth